+++ daml-lf/engine/src/main/scala/com/digitalasset/daml/lf/engine/Blinding.scala
@@ -134,6 +134,10 @@ object Blinding {
       }
     }
 
-    GenTransaction(roots = go(BackStack.empty, FrontStack(tx.roots)), nodes = filteredNodes)
+    GenTransaction(
+      roots = go(BackStack.empty, FrontStack(tx.roots)),
+      nodes = filteredNodes,
+      usedPackages = tx.usedPackages
+    )
   }
 }