+++ docs/source/support/release-notes.rst
@@ -14,6 +14,7 @@ HEAD — ongoing
   to evolve.
 - Remove DAML-LF Dev major version, ``--target dev`` option, and sandbox ``--allow-dev``
   option.  A "1.dev" target will handle the intended "Dev" use cases in a future release.
+- Include list of DAML packages used during interpretation in the produced transaction.
 
 0.12.11 - 2019-04-26
 --------------------